% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseExperimentData.R
\name{parseExperimentData}
\alias{parseExperimentData}
\title{parseExperimentData}
\usage{
parseExperimentData(data, lookupTable, numMeasurements = 3, numWells = 96)
}
\arguments{
\item{data}{A data frame containing raw experiment data in wide form}

\item{lookupTable}{A data frame containing per-well categorical fields (e.g. medium, clone, etc.)}

\item{numMeasurements}{(OPTIONAL) The number of measurements performed on the set of wells
If no argument is specified, numMeasurements defaults to 3.}

\item{numWells}{(OPTIONAL) The number of wells used in the experiment. If no argument is specified,
 numWells defaults to 96.}
}
\description{
Transform experiment data into long form with relevant keys
}
\details{
parseExperimentData will take raw experiment data in wide form and transform it into
 long form with the following keys:
 \itemize{
   \item{well}
   \item{measurementType}
   \item{cycle}
   \item{time}
 }
 The long-form experiment data is then joined to a lookup table containing qualitative information
 about each well. The qualitative data is replicated per-well for each instance of the well in
 the long-form data.

 parseExperimentData takes optional arguments specifying the number of measurements and the
 number of wells. This information is used to dynamically ascertain the measurement type for each
 iteration in the data and is appended as a new column. If no values are specified, it assumes
 3 measurements and 96 wells.
}

